Correct cross-platform Printf formatting of size_t item.

+++ plugins/dashboard_pi/src/dashboard_pi.cpp
@@ -1234,16 +1234,16 @@ bool dashboard_pi::SaveConfig( void )
         pConf->Write( _T("WindSpeedUnit"), g_iDashWindSpeedUnit );
 
         pConf->Write( _T("DashboardCount" ), (int) m_ArrayOfDashboardWindow.GetCount() );
-        for( size_t i = 0; i < m_ArrayOfDashboardWindow.GetCount(); i++ ) {
+        for( unsigned int i = 0; i < m_ArrayOfDashboardWindow.GetCount(); i++ ) {
             DashboardWindowContainer *cont = m_ArrayOfDashboardWindow.Item( i );
-            pConf->SetPath( wxString::Format( _T("/PlugIns/Dashboard/Dashboard%zu"), i + 1 ) );
+            pConf->SetPath( wxString::Format( _T("/PlugIns/Dashboard/Dashboard%d"), i + 1 ) );
             pConf->Write( _T("Name"), cont->m_sName );
             pConf->Write( _T("Caption"), cont->m_sCaption );
             pConf->Write( _T("Orientation"), cont->m_sOrientation );
 
             pConf->Write( _T("InstrumentCount"), (int) cont->m_aInstrumentList.GetCount() );
-            for( size_t j = 0; j < cont->m_aInstrumentList.GetCount(); j++ )
-                pConf->Write( wxString::Format( _T("Instrument%zu"), j + 1 ),
+            for( unsigned int j = 0; j < cont->m_aInstrumentList.GetCount(); j++ )
+                pConf->Write( wxString::Format( _T("Instrument%d"), j + 1 ),
                         cont->m_aInstrumentList.Item( j ) );
         }
